Move NounTile to common.
