+++ lib/widgets/nouns_screen/favorite_nouns_screen.dart
@@ -4,8 +4,8 @@ import 'package:provider/provider.dart';
 import 'package:elola/localizations.dart';
 import 'package:elola/configs/app_colors.dart';
 import 'package:elola/modules/noun_database/noun_database.dart';
-import 'package:elola/widgets/nouns_screen/noun_tile.dart';
 import 'package:elola/modules/player_data/player_data.dart';
+import 'package:elola/widgets/common/tiles/noun_tile.dart';
 
 class FavoriteNounsScreen extends StatelessWidget {
   const FavoriteNounsScreen({Key key}) : super(key: key);
