+++ lib/widgets/nouns_screen/nouns_screen.dart
@@ -4,7 +4,7 @@ import 'package:provider/provider.dart';
 import 'package:elola/enums/category.dart';
 import 'package:elola/models/noun.dart';
 import 'package:elola/modules/noun_database/noun_database.dart';
-import 'package:elola/widgets/nouns_screen/noun_tile.dart';
+import 'package:elola/widgets/common/tiles/noun_tile.dart';
 
 class NounsScreen extends StatelessWidget {
   final String title;
